+++ test/a.test.ts
@@ -1,72 +1,35 @@
-// // test.test.ts
-// import { describe, it, expect } from 'bun:test'
-// import { Elysia } from 'elysia'
+import { Elysia } from '../src'
+import { it, expect } from 'bun:test'
 
-// const setApp = new Elysia()
-// 	.onRequest(({ set }) => {
-// 		set.headers['x-header'] = 'test'
-// 		set.status = 400
-// 	})
-// 	.get('/', 'yay')
-// 	.get('/func', () => 'yay')
+it('DELETE requests returns 400 for no reason', async () => {
+	const app = new Elysia()
+		// I am doing group parse: ['application/json'] becouse I dont want to paste parse: ['application/json'] on every API
+		.group(
+			'user',
+			{ parse: 'application/json', detail: { tags: ['Users'] } },
+			(group) =>
+				group
+					.delete('/:id', ({ params: { id } }) => {
+						return id
+					}, {
+						parse: 'none'
+					})
+					.get('/', () => ({
+						id: '123',
+						name: 'John Doe'
+					}))
+			// and 10 GETs more..
+		)
+		.listen(3000)
 
-// const responseApp = new Elysia()
-// 	.onRequest(() => {
-// 		return new Response('nope', {
-// 			status: 400,
-// 			headers: {
-// 				'x-header': 'test'
-// 			}
-// 		})
-// 	})
-// 	.get('/', 'yay')
-// 	.get('/func', () => 'yay')
+	const responseForGet = await app.handle(
+		new Request('http://localhost/user', { method: 'GET' })
+	)
+	expect(responseForGet.status).toBe(200)
 
-// async function expectResponse(
-// 	response: Response,
-// 	exp: { status: number; header: string; body: string }
-// ) {
-// 	expect(response.status).toBe(exp.status)
-// 	expect(response.headers.get('x-header')).toBe(exp.header)
-// 	expect(await response.text()).toBe(exp.body)
-// }
-
-// const setExp = {
-// 	status: 200,
-// 	header: 'test',
-// 	body: 'yay'
-// }
-
-// const responseExp = {
-// 	status: 400,
-// 	header: 'test',
-// 	body: 'nope'
-// }
-
-// describe('Elysia', () => {
-// 	it('setApp - /', async () => {
-// 		const response = await setApp.handle(new Request('http://localhost/'))
-// 		await expectResponse(response, setExp)
-// 	})
-
-// 	it('setApp - /func', async () => {
-// 		const response = await setApp.handle(
-// 			new Request('http://localhost/func')
-// 		)
-// 		await expectResponse(response, setExp)
-// 	})
-
-// 	// it('responseApp - /', async () => {
-// 	// 	const response = await responseApp.handle(
-// 	// 		new Request('http://localhost/')
-// 	// 	)
-// 	// 	await expectResponse(response, responseExp)
-// 	// })
-
-// 	// it('responseApp - /func', async () => {
-// 	// 	const response = await responseApp.handle(
-// 	// 		new Request('http://localhost/func')
-// 	// 	)
-// 	// 	await expectResponse(response, responseExp)
-// 	// })
-// })
+	const responseForDelete = await app.handle(
+		new Request('http://localhost/user/123', { method: 'DELETE' })
+	)
+	// 👇 this is weird
+	expect(responseForDelete.status).toBe(400)
+})