+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Bug fix:
 - [#1028](https://github.com/elysiajs/elysia/issues/1028) string | string[] query parameter, reference array
 - [#1120](https://github.com/elysiajs/elysia/issues/1120) cannot set multiple cookies when response is a file
 - [#1124](https://github.com/elysiajs/elysia/issues/1124) validate url encoded query
+- [#1158](https://github.com/elysiajs/elysia/issues/1158) prevent side-effect from guard merge
 - [#1162](https://github.com/elysiajs/elysia/issues/1162) handle encoded space in array query string
 - [#1274](https://github.com/elysiajs/elysia/issues/1274) support .use(undefined | false) for conditional plugin
 - [#1276](https://github.com/elysiajs/elysia/issues/1276) mapResponse with set inference produce invalid instruction